fix: galeri

+++ app/Controllers/GaleriController.php
@@ -29,6 +29,11 @@ class GaleriController extends BaseController
         return view('galeri/tambah');
     }
 
+    public function addVideo()
+    {
+        return view('galeri/tambah-video');
+    }
+
     public function save()
     {
         if (!$this->validate([
@@ -72,13 +77,56 @@ class GaleriController extends BaseController
         return redirect()->to('/admin/galeri');
     }
 
+    public function saveVideo()
+    {
+        if (!$this->validate([
+            'judul' => [
+                'rules' => 'required|is_unique[galeri.judul]',
+                'errors' => [
+                    'required' => 'Judul harus diisi.',
+                    'is_unique' => 'Judul sudah digunakan.'
+                ]
+            ],
+            'deskripsi' => [
+                'rules' => 'required',
+                'errors' => [
+                    'required' => 'Deskripsi harus diisi.'
+                ]
+            ],
+            'link' => [
+                'rules' => 'required',
+                'errors' => [
+                    'required' => 'Link short harus diisi.'
+                ]
+            ],
+        ])) {
+            session()->setFlashdata('validator', $this->validator->getErrors());
+            return redirect()->to('/admin/galeri/add-video')->withInput();
+        }
+
+        $this->galeriModel->save([
+            'judul' => $this->request->getPost('judul'),
+            'deskripsi' => $this->request->getPost('deskripsi'),
+            'foto' => str_replace("shorts", "embed", $this->request->getPost('link')),
+            'type' => "video"
+        ]);
+
+        session()->setFlashdata('pesan', 'Data berhasil disimpan.');
+        return redirect()->to('/admin/galeri');
+    }
+
     public function edit($id)
     {
         $data = [
             'value' => $this->galeriModel->find($id)
         ];
 
-        return view('galeri/edit', $data);
+        if($data['value']['type'] == "foto"){
+            return view('galeri/edit', $data);
+        }else{
+            return view('galeri/edit-video', $data);
+        }
+
     }
 
     public function update($id)
@@ -129,10 +177,50 @@ class GaleriController extends BaseController
         return redirect()->to('/admin/galeri');
     }
 
+    public function updateVideo($id)
+    {
+        if (!$this->validate([
+            'judul' => [
+                'rules' => 'required|is_unique[galeri.judul,id_galeri,' . $id . ']',
+                'errors' => [
+                    'required' => 'Judul harus diisi.',
+                    'is_unique' => 'Judul sudah digunakan.'
+                ]
+            ],
+            'deskripsi' => [
+                'rules' => 'required',
+                'errors' => [
+                    'required' => 'Deskripsi harus diisi.'
+                ]
+            ],
+            'link' => [
+                'rules' => 'required',
+                'errors' => [
+                    'required' => 'Link short harus diisi.'
+                ]
+            ],
+        ])) {
+            session()->setFlashdata('validator', $this->validator->getErrors());
+            return redirect()->to('/admin/galeri/edit/' . $id)->withInput();
+        }
+
+        $this->galeriModel->update($id, [
+            'judul' => $this->request->getPost('judul'),
+            'deskripsi' => $this->request->getPost('deskripsi'),
+            'foto' => str_replace("shorts", "embed", $this->request->getPost('link')),
+        ]);
+
+        session()->setFlashdata('pesan', 'Data berhasil diubah.');
+        return redirect()->to('/admin/galeri');
+    }
+
     public function delete($id)
     {
         $galeri = $this->galeriModel->find($id);
-        unlink('img/' . $galeri['foto']);
+        if ($galeri['type'] == 'foto') {
+            unlink('img/' . $galeri['foto']);
+        }
+
         $this->galeriModel->delete($id);
 
         session()->setFlashdata('hapus', 'Data berhasil dihapus.');
